requirements.txt
pandas
numpy
matplotlib
sqlalchemy
mysql-connector-python
jupyter
python-dotenv
